[feature] - add more builders for specific samplers and listeners (#42)
- separate builders helps to keep same keyword names but use them in different context

+++ src/main/groovy/net/simonix/dsl/jmeter/factory/assertion/DurationAssertionFactory.groovy
@@ -47,7 +47,7 @@ final class DurationAssertionFactory extends TestElementNodeFactory {
 
     void updateTestElementProperties(TestElement testElement, Object name, Object value, Map config) {
         String applyTo = config.applyTo
-        Long duration = readValue(value, config.duration)
+        Long duration = readValue(value, config.duration) as Long
 
         if (applyTo == 'all') {
             testElement.setScopeAll()
